moved pages folder to components
