+++ project/src/components/pages/add-review-screen/add-review-screen.tsx
@@ -1,4 +1,4 @@
-import Header from '../../components/header/header';
+import Header from '../../header/header';
 
 function AddReviewScreen(): JSX.Element {
   return (
