+++ project/src/components/pages/not-found-screen/not-found-screen.tsx
@@ -1,5 +1,5 @@
-import Footer from '../../components/footer/footer';
-import Header from '../../components/header/header';
+import Footer from '../../footer/footer';
+import Header from '../../header/header';
 
 function NotFoundScreen(): JSX.Element {
   return (
