+++ project/src/components/pages/player-screen/player-screen.tsx
